combining analytics and ga files into one, updating test ref
