+++ src/utilities/ga.ts
@@ -1,4 +1,18 @@
-import { sendEvent } from './analytics';
+import { event } from 'react-ga';
+
+interface AnalyticsEvent {
+  category: string;
+  action: string;
+  label?: string;
+  value?: number;
+}
+
+/*
+ * Will throw error unless analytics is initialized
+ */
+export const sendEvent = (eventPayload: AnalyticsEvent) => {
+  event(eventPayload);
+};
 
 // LinodeActionMenu.tsx
 export const sendLinodeActionEvent = () => {
